EMP-416 ommitted review comments

+++ build.sh
@@ -8,6 +8,14 @@
 
 set -eu
 
+# common directory variables
+PREFIX="${PREFIX:-/usr/local}"
+DATAROOTDIR="${PREFIX}/share"
+DATADIR="${DATAROOTDIR}"
+EXEC_PREFIX="${PREFIX}"
+LIBEXECDIR="${EXEC_PREFIX}/libexec"
+SYSCONFDIR="${SYSCONFDIR:-/etc}"
+
 ARCH="${ARCH:-armhf}"
 ARM_EMU_BIN="${ARM_EMU_BIN:-}"
 
@@ -16,20 +24,20 @@ ALPINE_REPO="${ALPINE_REPO:-http://dl-cdn.alpinelinux.org/alpine}"
 
 TOOLBOX_IMAGE="${TOOLBOX_IMAGE:-um-update_toolbox.xz.img}"
 
-CUR_DIR="$(pwd)"
-BUILD_DIR="${CUR_DIR}/.build_${ARCH}"
+SRC_DIR="$(pwd)"
+NAME_TEMPLATE_BUILD_DIR=".build_${ARCH}"
+BUILD_DIR="${BUILD_DIR:-${SRC_DIR}/${NAME_TEMPLATE_BUILD_DIR}}"
 ROOTFS_DIR="${BUILD_DIR}/rootfs"
 
-SYSTEM_UPDATE_DIR="/etc/system_update"
+SYSTEM_UPDATE_SCRIPT_DIR="${LIBEXECDIR}/jedi_system_update.d"
 UPDATE_ROOTFS_SOURCE="/mnt/update_rootfs_source"
 
 PACKAGES="blkid busybox e2fsprogs-extra f2fs-tools rsync sfdisk"
 
 # Debian package information
 PACKAGE_NAME="${PACKAGE_NAME:-um-update-toolbox}"
-INSTALL_DIR="${INSTALL_DIR:-/usr/share/${PACKAGE_NAME}}"
+INSTALL_DIR="${INSTALL_DIR:-${DATADIR}/${PACKAGE_NAME}}"
 RELEASE_VERSION="${RELEASE_VERSION:-9999.99.99}"
-DEB_PACKAGE="${PACKAGE_NAME}_${ARCH}-${RELEASE_VERSION}.deb"
 
 
 cleanup()
@@ -45,7 +53,9 @@ cleanup()
         exit 1
     fi
 
-    rm -rf "${BUILD_DIR}"
+    if [ -z "${BUILD_DIR##*${NAME_TEMPLATE_BUILD_DIR}*}" ]; then
+        rm -rf "${BUILD_DIR:?}"
+    fi
 }
 
 bootstrap_prepare()
@@ -77,7 +87,7 @@ bootstrap_unprepare()
 
 add_update_scripts()
 {
-    target_script_dir="${ROOTFS_DIR}${SYSTEM_UPDATE_DIR}.d"
+    target_script_dir="${ROOTFS_DIR}/${SYSTEM_UPDATE_SCRIPT_DIR}"
     if [ ! -d "${target_script_dir}" ]; then
         mkdir -p "${target_script_dir}"
     fi
@@ -87,17 +97,15 @@ add_update_scripts()
         mkdir -p "${target_system_executable_dir}"
     fi
 
-    local_script_dir="${CUR_DIR}/scripts"
+    local_script_dir="${SRC_DIR}/scripts"
     for script in "${local_script_dir}"/[0-9][0-9]_*.sh; do
         basename="${script##*/}"
         echo "Installing ${script} on '${target_script_dir}/${basename}'."
         cp "${script}" "${target_script_dir}/${basename}"
-        chmod +x "${target_script_dir}/${basename}"
     done
 
     entrypoint_script="start_update.sh"
     cp "${local_script_dir}/${entrypoint_script}" "${target_system_executable_dir}/${entrypoint_script}"
-    chmod +x "${target_system_executable_dir}/${entrypoint_script}"
 }
 
 create_update_mount_points()
@@ -110,8 +118,8 @@ create_update_mount_points()
 
 add_configuration_files()
 {
-    local_config_dir="${CUR_DIR}/config"
-    target_config_dir="${ROOTFS_DIR}/etc/system_update"
+    local_config_dir="${SRC_DIR}/config"
+    target_config_dir="${ROOTFS_DIR}/${SYSCONFDIR}/jedi_system_update"
 
     if [ ! -d "${target_config_dir}" ]; then
         mkdir -p "${target_config_dir}"
@@ -128,8 +136,8 @@ bootstrap_rootfs()
 {
     echo "Bootstrapping Alpine Linux rootfs in to '${ROOTFS_DIR}'."
 
-    mkdir -p "${ROOTFS_DIR}/etc/apk"
-    echo "${ALPINE_REPO}/${ALPINE_VERSION}/main" > "${ROOTFS_DIR}/etc/apk/repositories"
+    mkdir -p "${ROOTFS_DIR}/${SYSCONFDIR}/apk"
+    echo "${ALPINE_REPO}/${ALPINE_VERSION}/main" > "${ROOTFS_DIR}/${SYSCONFDIR}/apk/repositories"
 
     # Install rootfs with base applications
     # shellcheck disable=SC2086
@@ -158,15 +166,18 @@ compress_rootfs()
 
 create_debian_package()
 {
-    deb_dir="${BUILD_DIR}/debian_deb_build"
+    DEB_DIR="${BUILD_DIR}/debian_deb_build"
 
-    mkdir -p "${deb_dir}/DEBIAN"
-    RELEASE_VERSION="${RELEASE_VERSION}" PACKAGE_NAME="${PACKAGE_NAME}" envsubst "\${RELEASE_VERSION} \${PACKAGE_NAME}" < "${CUR_DIR}/debian/control.in" > "${deb_dir}/DEBIAN/control"
+    mkdir -p "${DEB_DIR}/DEBIAN"
+    sed -e 's/@ARCH@/'"${ARCH}"'/g' \
+        -e 's/@PACKAGE_NAME@/'"${PACKAGE_NAME}"'/g' \
+        -e 's/@RELEASE_VERSION@/'"${RELEASE_VERSION}"'/g' \
+        "${SRC_DIR}/debian/control.in" > "${DEB_DIR}/DEBIAN/control"
 
-    mkdir -p "${deb_dir}${INSTALL_DIR}"
-    cp "${BUILD_DIR}/${TOOLBOX_IMAGE}" "${deb_dir}${INSTALL_DIR}/"
+    mkdir -p "${DEB_DIR}/${INSTALL_DIR}"
+    cp "${BUILD_DIR}/${TOOLBOX_IMAGE}" "${DEB_DIR}/${INSTALL_DIR}/"
 
-    dpkg-deb --build "${deb_dir}" "${BUILD_DIR}/${DEB_PACKAGE}"
+    dpkg-deb --build "${DEB_DIR}" "${BUILD_DIR}/${PACKAGE_NAME}_${ARCH}-${RELEASE_VERSION}.deb"
 }
 
 usage()
