+++ test/update_files.sh
@@ -11,26 +11,33 @@ set -eu
 # shellcheck source=test/include/chroot_env.sh
 . "test/include/chroot_env.sh"
 
+# common directory variables
+PREFIX="${PREFIX:-/usr/local}"
+EXEC_PREFIX="${PREFIX}"
+LIBEXECDIR="${EXEC_PREFIX}/libexec"
+SYSCONFDIR="${SYSCONFDIR:-/etc}"
+
 ARM_EMU_BIN="${ARM_EMU_BIN:-}"
 
-CWD="$(pwd)"
+SRC_DIR="$(pwd)"
 
-SYSTEM_UPDATE_DIR="${SYSTEM_UPDATE_DIR:-/etc/system_update}"
+SYSTEM_UPDATE_CONF_DIR="${SYSTEM_UPDATE_CONF_DIR:-${SYSCONFDIR}/jedi_system_update}"
+SYSTEM_UPDATE_SCRIPT_DIR="${SYSTEM_UPDATE_SCRIPT_DIR:-${LIBEXECDIR}/jedi_system_update.d/}"
 UPDATE_EXCLUDE_LIST_FILE="jedi_update_exclude_list.txt"
 UPDATE_ROOTFS_SOURCE="/tmp/update_source"
 TARGET_STORAGE_DEVICE=""
 
 JEDI_PARTITION_TABLE_FILE_NAME="config/jedi_emmc_sfdisk.table"
-UPDATE_FILES_COMMAND="${SYSTEM_UPDATE_DIR}.d/50_update_files.sh"
+UPDATE_FILES_COMMAND="${SYSTEM_UPDATE_SCRIPT_DIR}/50_update_files.sh"
 
-ULTIMAKER_VERSION_FILE="/etc/ultimaker_version"
-DEBIAN_VERSION_FILE="/etc/debian_version"
+ULTIMAKER_VERSION_FILE="${SYSCONFDIR}/ultimaker_version"
+DEBIAN_VERSION_FILE="${SYSCONFDIR}/debian_version"
 
 STORAGE_DEVICE_IMG="storage_device.img"
 BYTES_PER_SECTOR="512"
 STORAGE_DEVICE_SIZE="7553024" # sectors, about 3.6 GiB
 
-TEST_UPDATE_ROOTFS_FILE="${CWD}/test/test_rootfs.tar.xz"
+TEST_UPDATE_ROOTFS_FILE="${SRC_DIR}/test/test_rootfs.tar.xz"
 TEST_OUTPUT_FILE="$(mktemp -d)/test_results_$(basename "${0%.sh}").txt"
 
 NAME_TEMPLATE_TOOLBOX="um-update-toolbox"
@@ -62,7 +69,7 @@ create_dummy_storage_device()
 
     echo "writing partition table:"
 
-    sfdisk "${STORAGE_DEVICE_IMG}" < "${CWD}/${JEDI_PARTITION_TABLE_FILE_NAME}"
+    sfdisk "${STORAGE_DEVICE_IMG}" < "${SRC_DIR}/${JEDI_PARTITION_TABLE_FILE_NAME}"
 
     echo "formatting partitions"
 
@@ -80,11 +87,10 @@ create_dummy_storage_device()
 setup()
 {
     toolbox_root_dir="$(mktemp -d -t "${NAME_TEMPLATE_WORKDIR}.XXXXXX")"
-
     setup_chroot_env "${toolbox_image}" "${toolbox_root_dir}"
 
-    mkdir -p "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}"
-    tar xvfJ "${TEST_UPDATE_ROOTFS_FILE}" -C "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}" \
+    mkdir -p "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}"
+    tar -xJvf "${TEST_UPDATE_ROOTFS_FILE}" -C "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}" \
         > /dev/null 2> /dev/null
 
     work_dir="$(mktemp -d -t "${NAME_TEMPLATE_TOOLBOX}.XXXXXX")"
@@ -102,11 +108,11 @@ teardown()
         TARGET_STORAGE_DEVICE=""
     fi
 
-    if [ -d "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}" ] && [ -z "${toolbox_root_dir##*um-update-toolbox*}" ]; then
-        rm -rf "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}"
+    if [ -d "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}" ] && [ -z "${toolbox_root_dir##*um-update-toolbox*}" ]; then
+        rm -rf "${toolbox_root_dir:?}${UPDATE_ROOTFS_SOURCE}"
     fi
 
-    cd "${CWD}"
+    cd "${SRC_DIR}"
 
     if [ -d "${work_dir}" ] && [ -z "${work_dir##*update_files_workdir*}" ]; then
         rm -r "${work_dir}"
@@ -123,7 +129,7 @@ failure_exit()
     echo "When finished, the following is needed to cleanup!"
     echo "  sudo sh -c '\\"
     echo "    losetup -d '${TARGET_STORAGE_DEVICE}' && \\"
-    echo "    rm -rf '${work_dir:?}/*' && \\"
+    echo "    rm -rf '${work_dir}/*' && \\"
 
     failure_exit_chroot_env
 
@@ -167,59 +173,69 @@ run_test()
 
 test_missing_argument_update_rootfs_source_nok()
 {
-    chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" -s '' -d "${TARGET_STORAGE_DEVICE}" || return 0
+    eval chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" -s '' -d "${TARGET_STORAGE_DEVICE}" || return 0
 }
 
 test_update_rootfs_source_not_a_directory_nok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=/tmp/not_existing_dir"
+    chroot_environment=" \
+        UPDATE_ROOTFS_SOURCE=/tmp/not_existing_dir \
+        TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE} \
+    "
 
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
+    eval "${chroot_environment}" chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
 }
 
 test_invalid_block_device_nok()
 {
     target_device="/dev/loop100"
-    chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" -s "${UPDATE_ROOTFS_SOURCE}" -d "${target_device}" || return 0
+
+    eval chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" -s "${UPDATE_ROOTFS_SOURCE}" -d "${target_device}" || return 0
 }
 
 test_no_debian_distribution_found_nok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE}"
+    chroot_environment=" \
+        TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE} \
+        UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE} \
+    "
 
-    unlink "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}${DEBIAN_VERSION_FILE}"
+    unlink "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}/${DEBIAN_VERSION_FILE}"
 
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
+    eval "${chroot_environment}" chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" || return 0
 }
 
 test_no_ultimaker_software_found_nok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE}"
+    chroot_environment=" \
+        TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE} \
+        UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE} \
+    "
 
-    unlink "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}${ULTIMAKER_VERSION_FILE}"
+    unlink "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}/${ULTIMAKER_VERSION_FILE}"
 
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
+    eval "${chroot_environment}" chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" || return 0
 }
 
 test_update_files_ok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE}"
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 1
+    chroot_environment=" \
+        TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE} \
+        UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE} \
+    "
+
+    eval "${chroot_environment}" chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" || return 1
 
     update_target="$(mktemp -d -t "tmp_update_target.XXXXXX")"
     mount -t auto -v "${TARGET_STORAGE_DEVICE}p2" "${update_target}" || return 1
 
-    rsync --exclude-from "${CWD}/config/${UPDATE_EXCLUDE_LIST_FILE}" -c -a -x --dry-run \
-        "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}/" "${update_target}/" || return 1
+    rsync --exclude-from "${SRC_DIR}/config/${UPDATE_EXCLUDE_LIST_FILE}" -c -a -x --dry-run \
+        "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}/" "${update_target}/" || return 1
 
     umount "${update_target}"
 
     if [ -z "${update_target##*tmp_update_target*}" ]; then
-        rm -rf "${update_target}"
+        rm -rf "${update_target:?}"
     fi
 }
